% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directoryInput.R
\name{choose.dir}
\alias{choose.dir}
\title{Choose a Folder Interactively (Mac OS X)}
\arguments{
\item{default}{which folder to show initially}

\item{caption}{the caption on the selection dialog}
}
\value{
A length one character vector, character NA if 'Cancel' was selected.
}
\description{
Display a folder selection dialog under Mac OS X
}
