% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/papSmear.R
\docType{data}
\name{CNN}
\alias{CNN}
\title{CNN}
\usage{
data(CNN)
}
\description{
Trained Convolution Neuronal Network on images of unique cervix cells
             The model is build with mxnet package using 2 convolution  and 2 fully connected layers
}
\details{
90 rounds, 20 images per round, 7 folders that countain 7 cells classes
}
\author{
Karim Mezhoud \email{kmezhoud@gmail.com}
}
\keyword{data}
