% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertImage.R
\name{convert_image}
\alias{convert_image}
\title{Convert image to jpeg or png}
\usage{
convert_image(inputFolder, outputFolder, type = ".jpeg")
}
\arguments{
\item{inputFolder}{Folder that content the subfolders of images with other extension that jpeg and png. 
Each subfolder corresponds to one class.}

\item{outputFolder}{The folder where will save cponverted images.}

\item{type}{the type of extension jpeg, png all supported extesion by imager package}
}
\value{
save images as jpeg (default) or png. Each folder corresponds to a classe.
}
\description{
Convert image to jpeg or png
}
\examples{
\dontrun{
convert_image("/extdata/img_data/")
}

}
