% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/images2matrix.R
\name{im_split_df}
\alias{im_split_df}
\title{Function to split image into x/y portion and arrange them into dataframe.}
\usage{
im_split_df(path, X, Y)
}
\arguments{
\item{path}{path to image file (jpeg, png, bmp)}

\item{X}{integer, split X times image by x axis}

\item{Y}{integer, split Y times image by y axis}
}
\value{
a list of a dataframe each row corresponds to one image. 
w and h , the wide and the height of originale image
}
\description{
Function to split image into x/y portion and arrange them into dataframe.
}
\examples{
\dontrun{
im_split_df("path/image.jpg", 4, 3)
}
}
