% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/images2matrix.R
\name{images2matrix}
\alias{images2matrix}
\title{Convert images into matrix. Each row corresponds to one image}
\usage{
images2matrix(inputFolder, w = 28, h = 28, class)
}
\arguments{
\item{inputFolder}{Folder that content supported image bmp, jpeg, bnp.}

\item{w}{The resized width of image (pixels).}

\item{h}{The resized length of image (pixels)}

\item{class}{class label as integer.}
}
\value{
a data frame. The first column is the label of each image. the other column are the value of the pixels.
}
\description{
Convert images into matrix. Each row corresponds to one image
}
\examples{
\dontrun{
images2matrix("folder/path/", w= 28, h = 28, class= 1)
}

}
