% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Samp_TrainTest.R
\name{merge_train_test}
\alias{merge_train_test}
\title{Merge training and testing datasets from a list of dataframes (classes)}
\usage{
merge_train_test(list.df, n_test)
}
\arguments{
\item{list.df}{list of train and test dataframes}

\item{n_test}{number of test samples for each class}
}
\value{
A list of merged classes of tain and test dataframes
}
\description{
Merge training and testing datasets from a list of dataframes (classes)
}
\examples{
\dontrun{
merge_train_test(list.df, 5)
}
}
