% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/papSmear.R
\name{papSmear}
\alias{papSmear}
\title{Launch papSmear in the default browser}
\usage{
papSmear()
}
\value{
web page of papSmear web browser
}
\description{
Launch papSmear in the default browser
}
\details{
See \url{https://papSmear.github.io/docs} for documentation and tutorials
}
\examples{
\dontrun{
library(papSmear)
papSmear()
}
}
