% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/images2matrix.R
\name{vec2img}
\alias{vec2img}
\title{Visualize image from vector (a row from dataframe)}
\usage{
vec2img(df, nrow, w, h)
}
\arguments{
\item{df}{A dataframe WITHOUT rownames. 
Each row contents the pixels values of one image (as.vector). 
the length of the vector must to be the same for all images.}

\item{nrow}{numeric, the number of row want to be plot.}

\item{w}{wide of the image in pixels}

\item{h}{height of the image in pixels}
}
\value{
plot image
}
\description{
Visualize image from vector (a row from dataframe)
}
\examples{
\dontrun{
mat <- matrix(rnorm(2500, mean=65, sd=4.58), nrow= 25, ncol = 100)
vec2img(mat, 3, 10,10)
}

}
